[Loom] Share C benchmark API support
Move LoomC handle ownership, allocator conversion, and result failure
reporting into a small benchmark support library. Compile and link
throughput benchmarks can now share the public API lifecycle without
making either harness own the other.
