+++ loom/binding/c/benchmark/BUILD.bazel
@@ -30,6 +30,17 @@ iree_c_embed_data(
 )
 
 loom_cc_library(
+    name = "benchmark_support",
+    testonly = True,
+    srcs = ["util/benchmark_support.cc"],
+    hdrs = ["util/benchmark_support.h"],
+    deps = [
+        "//loom/binding/c:loomc",
+        "//runtime/src/iree/base",
+    ],
+)
+
+loom_cc_library(
     name = "compile_pool_prototype",
     testonly = True,
     srcs = ["util/compile_pool_prototype.c"],
@@ -49,6 +60,7 @@ loom_cc_library(
     hdrs = ["compile_throughput_benchmark.h"],
     deps = [
         ":benchmark_kernels",
+        ":benchmark_support",
         ":compile_pool_prototype",
         "//loom/binding/c:loomc",
         "//runtime/src/iree/base",
