+++ loom/binding/c/benchmark/compile_throughput_benchmark.cc
@@ -6,7 +6,6 @@
 
 #include "loom/binding/c/benchmark/compile_throughput_benchmark.h"
 
-#include <algorithm>
 #include <cstring>
 #include <string>
 #include <thread>
@@ -27,38 +26,6 @@ static bool ShouldSkipWorkerCount(::benchmark::State& state,
     return true;
   }
   return false;
-}
-
-static std::string FormatStatus(iree_status_t status) {
-  char buffer[4096] = {0};
-  iree_host_size_t length = 0;
-  iree_status_format(status, sizeof(buffer), buffer, &length);
-  return std::string(buffer, std::min(length, sizeof(buffer) - 1));
-}
-
-static iree_status_t MakeResultFailureStatus(const loomc_result_t* result,
-                                             const char* operation) {
-  if (result == nullptr) {
-    return iree_make_status(IREE_STATUS_FAILED_PRECONDITION,
-                            "%s failed without producing a result", operation);
-  }
-
-  loomc_host_size_t diagnostic_count = loomc_result_diagnostic_count(result);
-  if (diagnostic_count == 0) {
-    return iree_make_status(IREE_STATUS_FAILED_PRECONDITION,
-                            "%s failed without diagnostics", operation);
-  }
-
-  const loomc_diagnostic_t* diagnostic = loomc_result_diagnostic_at(result, 0);
-  if (diagnostic == nullptr) {
-    return iree_make_status(IREE_STATUS_FAILED_PRECONDITION,
-                            "%s failed with an unreadable diagnostic",
-                            operation);
-  }
-  return iree_make_status(
-      IREE_STATUS_FAILED_PRECONDITION, "%s failed: %.*s: %.*s", operation,
-      (int)diagnostic->code.size, diagnostic->code.data,
-      (int)diagnostic->message.size, diagnostic->message.data);
 }
 
 static iree_status_t RunScenarioJob(void* user_data,
@@ -107,24 +74,6 @@ static iree_status_t CreateTextSourceFromViews(loomc_string_view_t identifier,
 }
 
 }  // namespace
-
-iree_allocator_t host_allocator() { return iree_allocator_system(); }
-
-loomc_allocator_t loom_allocator() {
-  return loomc_allocator_from_iree(host_allocator());
-}
-
-iree_status_t to_iree_status(loomc_status_t status) {
-  return iree_status_from_loomc(status);
-}
-
-iree_status_t RequireSucceededResult(const loomc_result_t* result,
-                                     const char* operation) {
-  if (result != nullptr && loomc_result_succeeded(result)) {
-    return iree_ok_status();
-  }
-  return MakeResultFailureStatus(result, operation);
-}
 
 const loomc_artifact_t* FindArtifact(const loomc_result_t* result,
                                      loomc_artifact_kind_t kind,
