+++ loom/binding/c/benchmark/compile_throughput_benchmark.h
@@ -14,6 +14,7 @@
 #include <vector>
 
 #include "iree/base/api.h"
+#include "loom/binding/c/benchmark/util/benchmark_support.h"
 #include "loomc/iree.h"
 
 namespace benchmark {
@@ -21,31 +22,6 @@ class State;
 }  // namespace benchmark
 
 namespace loomc::bench {
-
-template <typename T, void (*Release)(T*)>
-struct HandleDeleter {
-  void operator()(T* handle) const { Release(handle); }
-};
-
-template <typename T, void (*Release)(T*)>
-using HandlePtr = std::unique_ptr<T, HandleDeleter<T, Release>>;
-
-using CompilerPtr = HandlePtr<loomc_compiler_t, loomc_compiler_release>;
-using ContextPtr = HandlePtr<loomc_context_t, loomc_context_release>;
-using LinkIndexBuilderPtr =
-    HandlePtr<loomc_link_index_builder_t, loomc_link_index_builder_release>;
-using LinkIndexPtr = HandlePtr<loomc_link_index_t, loomc_link_index_release>;
-using LinkerPtr = HandlePtr<loomc_linker_t, loomc_linker_release>;
-using ModulePtr = HandlePtr<loomc_module_t, loomc_module_release>;
-using PassProgramPtr =
-    HandlePtr<loomc_pass_program_t, loomc_pass_program_release>;
-using ResultPtr = HandlePtr<loomc_result_t, loomc_result_release>;
-using SourcePtr = HandlePtr<loomc_source_t, loomc_source_release>;
-using TargetEnvironmentPtr =
-    HandlePtr<loomc_target_environment_t, loomc_target_environment_release>;
-using TargetProfilePtr =
-    HandlePtr<loomc_target_profile_t, loomc_target_profile_release>;
-using WorkspacePtr = HandlePtr<loomc_workspace_t, loomc_workspace_release>;
 
 struct WorkerSlot {
   // Invocation-local scratch workspace used by one compile-pool worker.
@@ -139,12 +115,6 @@ class TargetCompileScenario : public CompileScenario {
 using CompileScenarioFactory = std::unique_ptr<CompileScenario> (*)(
     const ::benchmark::State& state, const void* user_data);
 
-iree_allocator_t host_allocator();
-
-loomc_allocator_t loom_allocator();
-
-iree_status_t to_iree_status(loomc_status_t status);
-
 void RunCompileBenchmark(::benchmark::State& state,
                          CompileScenarioFactory factory, const void* user_data);
 
@@ -157,9 +127,6 @@ void RunCompileBenchmarkDirect(::benchmark::State& state,
 void RunCompileBenchmarkDirectCold(::benchmark::State& state,
                                    CompileScenarioFactory factory,
                                    const void* user_data);
-
-iree_status_t RequireSucceededResult(const loomc_result_t* result,
-                                     const char* operation);
 
 const loomc_artifact_t* FindArtifact(const loomc_result_t* result,
                                      loomc_artifact_kind_t kind,
